initial shapefile reader
