+++ spatial/include/spatial/core/functions/table.hpp
@@ -9,12 +9,16 @@ struct CoreTableFunctions {
 public:
 	static void Register(DatabaseInstance &db) {
 		RegisterOsmTableFunction(db);
+
+		// TODO: Move these
 		RegisterShapefileTableFunction(db);
+		RegisterShapefileMetaTableFunction(db);
 	}
 
 private:
 	static void RegisterOsmTableFunction(DatabaseInstance &db);
 	static void RegisterShapefileTableFunction(DatabaseInstance &db);
+	static void RegisterShapefileMetaTableFunction(DatabaseInstance &db);
 };
 
 } // namespace core
