+++ spatial/src/spatial/core/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_subdirectory(geometry)
 add_subdirectory(functions)
+add_subdirectory(io)
+
 set(EXTENSION_SOURCES
         ${EXTENSION_SOURCES}
         ${CMAKE_CURRENT_SOURCE_DIR}/module.cpp
