+++ spatial/src/spatial/core/functions/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_subdirectory(aggregate)
 add_subdirectory(scalar)
 add_subdirectory(cast)
-add_subdirectory(table)
 
 set(EXTENSION_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/common.cpp
