+++ spatial/src/spatial/core/io/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_subdirectory(osm)
+add_subdirectory(shapefile)
+
+set(EXTENSION_SOURCES
+        ${EXTENSION_SOURCES}
+        PARENT_SCOPE
+)
