+++ spatial/src/spatial/core/io/osm/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(EXTENSION_SOURCES
         ${EXTENSION_SOURCES}
         ${CMAKE_CURRENT_SOURCE_DIR}/st_read_osm.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/st_read_shp.cpp
         PARENT_SCOPE
 )
