+++ spatial/src/spatial/core/io/osm/st_read_osm.cpp
