+++ spatial/src/spatial/core/io/shapefile/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(EXTENSION_SOURCES
+        ${EXTENSION_SOURCES}
+        ${CMAKE_CURRENT_SOURCE_DIR}/read_shapefile.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/read_shapefile_meta.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/shapefile_common.cpp
+        PARENT_SCOPE
+)
